minor: remove unused crypto functions & narrow public API
